requirements.txt
django==3.0.3
djangorestframework==3.12.2
django-rest-swagger==2.2.0
django-debug-toolbar==3.2
django-filter==2.4.0
psycopg2-binary==2.8.4
Pillow==8.1.2
python-dotenv==0.15.0
